single class for Attribute upload
got rid of the auxilary class for attributes upload.
