+++ csharp/Urban.DCP.Data/Urban.DCP.Data.csproj
@@ -148,7 +148,6 @@
     <Compile Include="PropertyCommentInfo.cs" />
     <Compile Include="Tests\CommentTests.cs" />
     <Compile Include="UiComment.cs" />
-    <Compile Include="Uploadable\AttributeUploadable.cs" />
     <Compile Include="Uploadable\ImportResult.cs" />
     <Compile Include="Uploadable\Project.cs" />
     <Compile Include="ResultsWithMetadata.cs" />
